Move extracted components to separate files
